refactor(tests): Extract shared test doubles to conftest.py
- Consolidate MockHostConfig (4 copies), TEST_SIGNING_KEY (4 copies),
  and mint_test_token (4 implementations) into conftest.py
- Remove redundant sys.path.insert from 27 test files
- Net -183 lines while maintaining all 637 tests

Closes #277
