+++ tests/test_actions_ansible.py
@@ -3,22 +3,10 @@
 Tests for AnsiblePlaybookAction, AnsibleLocalPlaybookAction, and EnsurePVEAction.
 """
 
-import sys
-from dataclasses import dataclass
 from pathlib import Path
 from unittest.mock import patch, MagicMock
 
-sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
-
-
-@dataclass
-class MockHostConfig:
-    """Minimal host config for testing."""
-    name: str = 'test-host'
-    ssh_host: str = '192.0.2.1'
-    ssh_user: str = 'root'
-    automation_user: str = 'homestak'
-    config_file: Path = Path('/tmp/test.yaml')
+from conftest import MockHostConfig
 
 
 class TestAnsiblePlaybookAction:
