+++ tests/test_actions_file.py
@@ -3,11 +3,7 @@
 Tests for DownloadGitHubReleaseAction including split file handling.
 """
 
-import sys
-from pathlib import Path
 from unittest.mock import patch, MagicMock
-
-sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
 
 
 class TestDownloadGitHubReleaseAction:
