+++ tests/test_actions_ssh.py
@@ -4,24 +4,11 @@ Tests for SSHCommandAction, WaitForSSHAction, WaitForFileAction,
 VerifyPackagesAction, VerifyUserAction, and ActionResult.
 """
 
-import sys
-from dataclasses import dataclass
 from pathlib import Path
 from unittest.mock import patch
 
-sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
-
 from common import ActionResult
-
-
-@dataclass
-class MockHostConfig:
-    """Minimal host config for testing."""
-    name: str = 'test-host'
-    ssh_host: str = '192.0.2.1'  # TEST-NET-1 (RFC 5737)
-    ssh_user: str = 'root'  # For PVE host connections
-    automation_user: str = 'homestak'  # For VM connections
-    config_file: Path = Path('/tmp/test.yaml')
+from conftest import MockHostConfig
 
 
 class TestSSHCommandAction:
