+++ tests/test_actions_tofu.py
@@ -3,11 +3,7 @@
 Tests for TofuApplyAction and TofuDestroyAction.
 """
 
-import sys
-from pathlib import Path
 from unittest.mock import patch, MagicMock
-
-sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
 
 
 class TestTofuApplyAction:
