+++ tests/test_config.py
@@ -10,11 +10,8 @@ Tests verify:
 """
 
 import os
-import sys
 from pathlib import Path
 from unittest.mock import patch
-
-sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
 
 import pytest
 from config import (
