+++ tests/test_config_resolver.py
@@ -10,12 +10,8 @@ Tests verify:
 """
 
 import json
-import sys
-from pathlib import Path
 
 import yaml
-
-sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
 
 import pytest
 from config import ConfigError
