+++ tests/test_manifest.py
@@ -11,12 +11,9 @@ Tests verify:
 """
 
 import json
-import sys
 from pathlib import Path
 from unittest.mock import patch, MagicMock
 import tempfile
-
-sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
 
 import pytest
 from config import ConfigError
