+++ tests/test_manifest_validate.py
@@ -10,12 +10,8 @@ Tests verify:
 6. CLI exit codes
 """
 
-import sys
-from pathlib import Path
 from unittest.mock import patch
 import tempfile
-
-sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
 
 import pytest
 from manifest import Manifest, ManifestNode
