+++ tests/test_operator_graph.py
@@ -1,12 +1,6 @@
 """Tests for manifest_opr.graph module."""
 
-import sys
-from pathlib import Path
-
 import pytest
-
-# Add src to path for imports
-sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
 
 from manifest import Manifest, ManifestNode
 from manifest_opr.graph import ExecutionNode, ManifestGraph
