+++ tests/test_readiness.py
@@ -1,11 +1,7 @@
 #!/usr/bin/env python3
 """Tests for readiness checks."""
 
-import sys
-from pathlib import Path
 from unittest.mock import patch, MagicMock
-
-sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
 
 import pytest
 from readiness import (
