+++ tests/test_scenario_attributes.py
@@ -7,12 +7,6 @@ These tests verify that:
 3. CLI getattr pattern works as expected
 """
 
-import sys
-from pathlib import Path
-
-# Add src to path
-sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
-
 import pytest
 from scenarios import get_scenario, list_scenarios
